backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-extensions==3.1.3
django-redis==4.12.1
djangorestframework==3.12.4
psycopg2==2.9.1
pyaml==20.4.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
sqlparse==0.4.1
